Add generic vars support
vars can now be specific in the _distro_map.yml to pass in
arbitrary distro or branch specific values to the template
processor.  They show up in the distro_vars and branch_vars
variables.  The name is lower case to match the others in
_distro_map.yml

Vars can be specified (upper case because other values are
upper case in _topic_map.yml) can be specified for both Dir
and File entries.  They show up in the template in the variable
topic_vars.  The vars are consolidated so you get all variables
from the top to what ever topic you are processing.  Lower level
vars override those from the top.